docker/requirements.txt
azure-storage-queue
requests
pymysql